Implementação das Validações

+++ app/Http/Controllers/CertificacaoController.php
@@ -18,7 +18,7 @@ class CertificacaoController extends Controller
     public function index()
     {
         $certificacoes = $this->certificacao->all();
-        return $certificacoes;
+        return response()->json($certificacoes, 200);
     }
 
     /**
@@ -30,8 +30,9 @@ class CertificacaoController extends Controller
     public function store(Request $request)
     {
         // $certificacao = Certificacao::create($request->all());
+        $request->validate($this->certificacao->rules(), $this->certificacao->feedback());
         $certificacao = $this->certificacao->create($request->all());
-        return $certificacao;
+        return response()->json($certificacao, 201);
     }
 
     /**
@@ -43,7 +44,10 @@ class CertificacaoController extends Controller
     public function show($id)
     {
         $certificacao = $this->certificacao->find($id);
-        return $certificacao;
+        if($certificacao === null) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe'], 404);
+        } 
+        return response()->json($certificacao, 200);
     }
 
     /**
@@ -57,8 +61,32 @@ class CertificacaoController extends Controller
     {
         // $certificacao->update($request->all());
         $certificacao = $this->certificacao->find($id);
+
+        if($certificacao === null) {
+            return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'], 404);
+        }
+
+        if($request->method() === 'PATCH') {
+
+            $regrasDinamicas = array();
+
+            //percorrendo todas as regras definidas no Model
+            foreach($certificacao->rules() as $input => $regra) {
+                
+                //coletar apenas as regras aplicáveis aos parâmetros parciais da requisição PATCH
+                if(array_key_exists($input, $request->all())) {
+                    $regrasDinamicas[$input] = $regra;
+                }
+            }
+            
+            $request->validate($regrasDinamicas, $certificacao->feedback());
+
+        } else {
+            $request->validate($certificacao->rules(), $certificacao->feedback());
+        }
+
         $certificacao->update($request->all());
-        return $certificacao;
+        return response()->json($certificacao, 200);
     }
 
     /**
@@ -71,7 +99,12 @@ class CertificacaoController extends Controller
     {
         // $certificacao->delete();
         $certificacao = $this->certificacao->find($id);
+
+        if($certificacao === null) {
+            return response()->json(['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'], 404);
+        }
+
         $certificacao->delete();
-        return ['msg' => 'O Certificado foi removido com sucesso!'];
+        return response()->json(['msg' => 'O Certificado foi removido com sucesso!'], 200);
     }
 }
