+++ app/Http/Controllers/InteresseController.php
@@ -20,7 +20,7 @@ class InteresseController extends Controller
     {
         // $interesses = Interesse::all();
         $interesses = $this->interesse->all();
-        return $interesses;
+        return response()->json($interesses,200);
     }
 
     /**
@@ -32,8 +32,9 @@ class InteresseController extends Controller
     public function store(Request $request)
     {
         // $interesse = Interesse::create($request->all());
+        $request->validate($this->interesse->rules(), $this->interesse->feedback());
         $interesse = $this->interesse->create($request->all());
-        return $interesse;
+        return response()->json($interesse,200);
     }
 
     /**
@@ -45,6 +46,11 @@ class InteresseController extends Controller
     public function show($id)
     {
         $interesse = $this->interesse->find($id);
+
+        if($interesse === null) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe'], 404);
+        }
+
         return $interesse;
     }
 
@@ -59,8 +65,32 @@ class InteresseController extends Controller
     {
         // $interesse->update($request->all());
         $interesse = $this->interesse->find($id);
+
+        if($interesse === null) {
+            return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'], 404);
+        }
+
+        if($request->method() === 'PATCH') {
+
+            $regrasDinamicas = array();
+
+            //percorrendo todas as regras definidas no Model
+            foreach($interesse->rules() as $input => $regra) {
+                
+                //coletar apenas as regras aplicáveis aos parâmetros parciais da requisição PATCH
+                if(array_key_exists($input, $request->all())) {
+                    $regrasDinamicas[$input] = $regra;
+                }
+            }
+            
+            $request->validate($regrasDinamicas, $interesse->feedback());
+
+        } else {
+            $request->validate($interesse->rules(), $interesse->feedback());
+        }
+
         $interesse->update($request->all());
-        return $interesse;
+        return response()->json($interesse,200);
     }
 
     /**
@@ -73,7 +103,12 @@ class InteresseController extends Controller
     {
         // $interesse->delete();
         $interesse = $this->interesse->find($id);
+
+        if($interesse === null) {
+            return response()->json(['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'], 404);
+        }
+
         $interesse->delete();
-        return ['msg' => 'O interesse foi removido com sucesso!'];
+        return response()->json(['msg' => 'O interesse foi removido com sucesso!'],200);
     }
 }
