+++ app/Http/Controllers/PessoaController.php
@@ -20,7 +20,7 @@ class PessoaController extends Controller
     {
         // $pessoas = Pessoa::all();
         $pessoas = $this->pessoa->all();
-        return $pessoas;
+        return response()->json($pessoas,200);
     }
 
     /**
@@ -32,8 +32,9 @@ class PessoaController extends Controller
     public function store(Request $request)
     {
         // $pessoa = Pessoa::create($request->all());
+        $request->validate($this->pessoa->rules(), $this->pessoa->feedback());
         $pessoa = $this->pessoa->create($request->all());
-        return $pessoa;
+        return response()->json($pessoa,200);
     }
 
     /**
@@ -45,7 +46,12 @@ class PessoaController extends Controller
     public function show($id)
     {
         $pessoa = $this->pessoa->find($id);
-        return $pessoa;
+
+        if($pessoa === null) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe'], 404);
+        }
+
+        return response()->json($pessoa,200);
     }
 
     /**
@@ -59,8 +65,32 @@ class PessoaController extends Controller
     {
         // $pessoa->update($request->all());
         $pessoa = $this->pessoa->find($id);
+
+        if($pessoa === null) {
+            return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'], 404);
+        }
+
+        if($request->method() === 'PATCH') {
+
+            $regrasDinamicas = array();
+
+            //percorrendo todas as regras definidas no Model
+            foreach($pessoa->rules() as $input => $regra) {
+                
+                //coletar apenas as regras aplicáveis aos parâmetros parciais da requisição PATCH
+                if(array_key_exists($input, $request->all())) {
+                    $regrasDinamicas[$input] = $regra;
+                }
+            }
+            
+            $request->validate($regrasDinamicas, $pessoa->feedback());
+
+        } else {
+            $request->validate($pessoa->rules(), $pessoa->feedback());
+        }
+
         $pessoa->update($request->all());
-        return $pessoa;
+        return response()->json($pessoa,200);
     }
 
     /**
@@ -73,7 +103,12 @@ class PessoaController extends Controller
     {
         // $pessoa->delete();
         $pessoa = $this->pessoa->find($id);
+
+        if($pessoa === null) {
+            return response()->json(['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'], 404);
+        }
+
         $pessoa->delete();
-        return ['msg' => 'A pessoa foi removida com sucesso!'];
+        return response()->json(['msg' => 'A pessoa foi removida com sucesso!'],200);
     }
 }
