+++ app/Http/Controllers/SkillController.php
@@ -20,7 +20,7 @@ class SkillController extends Controller
     {
         // $skills = SKILL::all();
         $skills = $this->skills->all();
-        return $skills;
+        return response()->json($skills,200);
     }
 
     /**
@@ -32,8 +32,9 @@ class SkillController extends Controller
     public function store(Request $request)
     {
         // $skill = Skill::create($request->all());
+        $request->validate($this->skill->rules(), $this->skill->feedback());
         $skill = $this->skill->create($request->all());
-        return $skill;
+        return response()->json($skill,200);
     }
 
     /**
@@ -45,7 +46,12 @@ class SkillController extends Controller
     public function show($id)
     {
         $skill = $this->skill->find($id);
-        return $skill;
+
+        if($skill === null) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe'], 404);
+        }
+
+        return response()->json($skill,200);
     }
 
     /**
@@ -59,8 +65,32 @@ class SkillController extends Controller
     {
         // $skill->update($request->all());
         $skill = $this->skill->find($id);
+
+        if($skill === null) {
+            return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'], 404);
+        }
+
+        if($request->method() === 'PATCH') {
+
+            $regrasDinamicas = array();
+
+            //percorrendo todas as regras definidas no Model
+            foreach($skill->rules() as $input => $regra) {
+                
+                //coletar apenas as regras aplicáveis aos parâmetros parciais da requisição PATCH
+                if(array_key_exists($input, $request->all())) {
+                    $regrasDinamicas[$input] = $regra;
+                }
+            }
+            
+            $request->validate($regrasDinamicas, $skill->feedback());
+
+        } else {
+            $request->validate($skill->rules(), $skill->feedback());
+        }
+
         $skill->update($request->all());
-        return $skill;
+        return response()->json($skill,200);
     }
 
     /**
@@ -73,7 +103,12 @@ class SkillController extends Controller
     {
         // $skill->delete();
         $skill = $this->skill->find($id);
+
+        if($skill === null) {
+            return response()->json(['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'], 404);
+        }
+
         $skill->delete();
-        return ['msg' => 'O skill foi removido com sucesso!'];
+        return response()->json(['msg' => 'O skill foi removido com sucesso!'],200);
     }
 }
